requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
Pillow
django-filter>=23.5
djangorestframework-simplejwt
reportlab
django-cors-headers